move repository call in Service
